Refactor ubrn_bindgen cli
