+++ crates/ubrn_bindgen/src/lib.rs
@@ -4,6 +4,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
 mod bindings;
+mod cli;
+mod switches;
 
-pub use bindings::metadata::ModuleMetadata;
-pub use bindings::{AbiFlavor, BindingsArgs, OutputArgs, SourceArgs, SwitchArgs};
+pub use self::{
+    bindings::{gen_cpp::generate_entrypoint, metadata::ModuleMetadata},
+    cli::{BindingsArgs, OutputArgs, SourceArgs},
+    switches::{AbiFlavor, SwitchArgs},
+};
